ref: more @mittorn's changes on RefAPI
