+++ ref_gl/gl_context.c
@@ -12,8 +12,7 @@ but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
-#include "common.h"
-#include "client.h"
+
 #include "gl_local.h"
 #include "gl_export.h"
