+++ ref_gl/gl_frustum.c
@@ -13,7 +13,6 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "common.h"
 #include "gl_local.h"
 #include "mathlib.h"
