+++ ref_gl/gl_image.c
@@ -13,8 +13,6 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "common.h"
-#include "client.h"
 #include "gl_local.h"
 
 #define TEXTURES_HASH_SIZE	(MAX_TEXTURES >> 2)
