+++ ref_gl/gl_rmath.c
@@ -13,10 +13,8 @@ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 */
 
-#include "common.h"
 #include "gl_local.h"
 #include "mathlib.h"
-#include "client.h"
 
 /*
 ========================================================================
