+++ ref_gl/wscript
@@ -33,7 +33,8 @@ def build(bld):
 	source = bld.path.ant_glob(['*.c'])
 
 	includes = ['.',
-	    '../engine/common',
+		'../engine',
+		'../engine/common',
 		'../engine/server',
 		'../engine/client',
 		'../common',
